json parser test coverage and refactor
