+++ src/test/java/io/nats/client/support/JsonUtilsTests.java
@@ -28,10 +28,7 @@ import java.util.concurrent.atomic.AtomicLong;
 import java.util.regex.Pattern;
 
 import static io.nats.client.support.ApiConstants.DESCRIPTION;
-import static io.nats.client.support.Encoding.jsonDecode;
-import static io.nats.client.support.Encoding.jsonEncode;
 import static io.nats.client.support.JsonUtils.*;
-import static io.nats.client.utils.ResourceUtils.dataAsLines;
 import static io.nats.client.utils.ResourceUtils.dataAsString;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -373,52 +370,6 @@ public final class JsonUtilsTests {
     }
 
     @Test
-    public void testEncodeDecode() {
-        _testEncodeDecode("b4\\\\after", "b4\\after", null); // a single slash with a meaningless letter after it
-        _testEncodeDecode("b4\\\\tafter", "b4\\tafter", null); // a single slash with a char that can be part of an escape
-
-        _testEncodeDecode("b4\\bafter", "b4\bafter", null);
-        _testEncodeDecode("b4\\fafter", "b4\fafter", null);
-        _testEncodeDecode("b4\\nafter", "b4\nafter", null);
-        _testEncodeDecode("b4\\rafter", "b4\rafter", null);
-        _testEncodeDecode("b4\\tafter", "b4\tafter", null);
-
-        _testEncodeDecode("b4\\u0000after", "b4" + (char) 0 + "after", null);
-        _testEncodeDecode("b4\\u001fafter", "b4" + (char) 0x1f + "after", "b4\\u001fafter");
-        _testEncodeDecode("b4\\u0020after", "b4 after", "b4 after");
-        _testEncodeDecode("b4\\u0022after", "b4\"after", "b4\\\"after");
-        _testEncodeDecode("b4\\u0027after", "b4'after", "b4'after");
-        _testEncodeDecode("b4\\u003dafter", "b4=after", "b4=after");
-        _testEncodeDecode("b4\\u003Dafter", "b4=after", "b4=after");
-        _testEncodeDecode("b4\\u003cafter", "b4<after", "b4<after");
-        _testEncodeDecode("b4\\u003Cafter", "b4<after", "b4<after");
-        _testEncodeDecode("b4\\u003eafter", "b4>after", "b4>after");
-        _testEncodeDecode("b4\\u003Eafter", "b4>after", "b4>after");
-        _testEncodeDecode("b4\\u0060after", "b4`after", "b4`after");
-        _testEncodeDecode("b4\\xafter", "b4xafter", "b4xafter"); // unknown escape
-        _testEncodeDecode("b4\\", "b4\\", "b4\\\\"); // last char is \
-        _testEncodeDecode("b4\\/after", "b4/after", null);
-
-        List<String> utfs = dataAsLines("utf8-only-no-ws-test-strings.txt");
-        for (String u : utfs) {
-            String uu = "b4\\b\\f\\n\\r\\t" + u + "after";
-            _testEncodeDecode(uu, "b4\b\f\n\r\t" + u + "after", uu);
-        }
-    }
-
-    private void _testEncodeDecode(String encodedInput, String targetDecode, String targetEncode) {
-        String decoded = jsonDecode(encodedInput);
-        assertEquals(targetDecode, decoded);
-        String encoded = jsonEncode(new StringBuilder(), decoded).toString();
-        if (targetEncode == null) {
-            assertEquals(encodedInput, encoded);
-        }
-        else {
-            assertEquals(targetEncode, encoded);
-        }
-    }
-
-    @Test
     public void testSimpleMessageBody() {
         assertEquals("{\"number\":1}", new String(simpleMessageBody("number", 1)));
         assertEquals("{\"string\":\"str\"}", new String(simpleMessageBody("string", "str")));
@@ -427,6 +378,7 @@ public final class JsonUtilsTests {
     @Test
     public void testMiscCoverage() {
         Pattern ipattern = integer_pattern("foo");
+        //noinspection deprecation
         Pattern npattern = number_pattern("foo"); // coverage for deprecated
 
         assertEquals(ipattern.pattern(), npattern.pattern());
